api/cs415/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
PyJWT==2.8.0
typing_extensions==4.9.0